refactor(manager): split manager.py into mixin modules (A2)
Extracts RefreshMixin (_refresh.py) and PersistenceMixin (_persistence.py)
from the 1321-line FinanceDashboardManager into focused single-responsibility
modules. The main class in manager/__init__.py inherits both mixins and retains
lifecycle, demo, balance/summary reads, household, and transfer-chain logic.

MRO: FinanceDashboardManager → RefreshMixin → PersistenceMixin → object.

The external import path 'from .manager import FinanceDashboardManager' is
unchanged. All 49 existing tests pass.
